+++ src/Effects/ParticlesSL/Particle_Stuff/particleUtilsPS.cpp
@@ -12,7 +12,7 @@ void particleUtilsPS::resetParticle(particlePS *particle) {
 }
 
 //resets a given particle in a particle set back to it's starting position and sets it's update time to 0
-void particleUtilsPS::resetParticle(particleSetPS &particleSet, uint8_t partNum) {
+void particleUtilsPS::resetParticle(particleSetPS &particleSet, uint16_t partNum) {
     resetParticle(particleSet.particleArr[partNum]);
 }
 
@@ -29,7 +29,7 @@ void particleUtilsPS::resetParticleSet(particleSetPS &particleSet) {
 //See each of the individual functions for what the options do
 //NOTE: the particles are created using new, so don't forget to call freeParticleSet() or freeParticle() once you are
 //finished with a particleSet
-particleSetPS particleUtilsPS::buildParticleSet(uint8_t numParticles, uint16_t maxPosition, uint8_t direction, uint16_t baseSpeed,
+particleSetPS particleUtilsPS::buildParticleSet(uint16_t numParticles, uint16_t maxPosition, uint8_t direction, uint16_t baseSpeed,
                                                 uint16_t speedRange, uint16_t size, uint16_t sizeRange, uint8_t trailType,
                                                 uint8_t trailSize, uint8_t trailRange, uint8_t bounce, uint8_t colorIndex,
                                                 bool randColor) {
@@ -39,7 +39,7 @@ particleSetPS particleUtilsPS::buildParticleSet(uint8_t numParticles, uint16_t m
     particleSetPS newParticleSet = {particleArr, numParticles, numParticles};
 
     //create a new set of particles
-    for( uint8_t i = 0; i < numParticles; i++ ) {
+    for( uint16_t i = 0; i < numParticles; i++ ) {
         //particlePS *p = new particlePS();
         particlePS *p = (particlePS *)malloc(sizeof(particlePS));
         newParticleSet.setParticle(p, i);
@@ -62,7 +62,7 @@ particleSetPS particleUtilsPS::buildParticleSet(uint8_t numParticles, uint16_t m
 void particleUtilsPS::randomizeParticleSet(particleSetPS &particleSet, uint16_t maxPosition, uint8_t direction, uint16_t baseSpeed,
                                            uint16_t speedRange, uint16_t size, uint16_t sizeRange, uint8_t trailType, uint8_t trailSize,
                                            uint8_t trailRange, uint8_t bounce, uint8_t colorIndex, bool randColor) {
-    for( uint8_t i = 0; i < particleSet.length; i++ ) {
+    for( uint16_t i = 0; i < particleSet.length; i++ ) {
         randomizeParticle(particleSet, i, maxPosition, direction, baseSpeed, speedRange, size,
                           sizeRange, trailType, trailSize, trailRange, bounce, colorIndex, randColor);
     }
@@ -72,7 +72,7 @@ void particleUtilsPS::randomizeParticleSet(particleSetPS &particleSet, uint16_t 
 //The option names are the same as used in the individual setParticle functions and do the same things
 //ie maxPosition will be passed to setParticleSetPositions() as the maxPosition input
 //See each of the individual functions for what the options do
-void particleUtilsPS::randomizeParticle(particleSetPS &particleSet, uint8_t partNum, uint16_t maxPosition, uint8_t direction, uint16_t baseSpeed,
+void particleUtilsPS::randomizeParticle(particleSetPS &particleSet, uint16_t partNum, uint16_t maxPosition, uint8_t direction, uint16_t baseSpeed,
                                         uint16_t speedRange, uint16_t size, uint16_t sizeRange, uint8_t trailType, uint8_t trailSize,
                                         uint8_t trailRange, uint8_t bounce, uint8_t colorIndex, bool randColor) {
 
@@ -111,7 +111,7 @@ void particleUtilsPS::setParticleSetProp(particleSetPS &particleSet, uint8_t pro
     if( opt2 != 0 ) {
         randColor = true;
     }
-    for( uint8_t i = 0; i < particleSetLength; i++ ) {
+    for( uint16_t i = 0; i < particleSetLength; i++ ) {
         switch( propNum ) {
             case 0:
                 setParticleSetPosition(particleSet, i, opt1, opt2);
@@ -143,7 +143,7 @@ void particleUtilsPS::setParticleSetProp(particleSetPS &particleSet, uint8_t pro
 //Sets a particle's startPosition to the passed in value
 //if rand is true, it will be chosen randomly up to the passed in position
 //(generally use the SegmentSetPS length as the position for rand)
-void particleUtilsPS::setParticleSetPosition(particleSetPS &particleSet, uint8_t partNum, uint16_t position, bool rand) {
+void particleUtilsPS::setParticleSetPosition(particleSetPS &particleSet, uint16_t partNum, uint16_t position, bool rand) {
     if( rand ) {
         position = random16(position);
     }
@@ -152,7 +152,7 @@ void particleUtilsPS::setParticleSetPosition(particleSetPS &particleSet, uint8_t
 
 //sets a particle's direction based on the passed in direction
 //to set the direction randomly, pass in a value > 1
-void particleUtilsPS::setParticleSetDirection(particleSetPS &particleSet, uint8_t partNum, uint8_t direction) {
+void particleUtilsPS::setParticleSetDirection(particleSetPS &particleSet, uint16_t partNum, uint8_t direction) {
     if( direction > 1 ) {
         direction = random8(2);
     }
@@ -162,13 +162,13 @@ void particleUtilsPS::setParticleSetDirection(particleSetPS &particleSet, uint8_
 //sets a particle's speed to the passed in baseSpeed (ms)
 //The speed is varied by a random amount no greater than the passed in range
 //(the baseSpeed is the fastest a particle will go)
-void particleUtilsPS::setParticleSetSpeed(particleSetPS &particleSet, uint8_t partNum, uint16_t baseSpeed, uint16_t range) {
+void particleUtilsPS::setParticleSetSpeed(particleSetPS &particleSet, uint16_t partNum, uint16_t baseSpeed, uint16_t range) {
     particleSet.particleArr[partNum]->speed = baseSpeed + random16(range + 1);
 }
 
 //sets a particle's size to the passed in size
 //The size is varied by a random amount no greater than the passed in range
-void particleUtilsPS::setParticleSetSize(particleSetPS &particleSet, uint8_t partNum, uint16_t size, uint16_t range) {
+void particleUtilsPS::setParticleSetSize(particleSetPS &particleSet, uint16_t partNum, uint16_t size, uint16_t range) {
     //particles must have a minimum size of 1
     if( size == 0 ) {
         size = 1;
@@ -180,7 +180,7 @@ void particleUtilsPS::setParticleSetSize(particleSetPS &particleSet, uint8_t par
 //if you pass in a trailType > 4 (the highest trail type) the trails will be chosen randomly from the
 //first three trail types (0, 1, 2)
 //If you want a specific set of trail types chosen randomly, use setParticleSetTrailRand()
-void particleUtilsPS::setParticleSetTrailType(particleSetPS &particleSet, uint8_t partNum, uint8_t trailType) {
+void particleUtilsPS::setParticleSetTrailType(particleSetPS &particleSet, uint16_t partNum, uint8_t trailType) {
     if( trailType > 4 ) {
         trailType = random8(3);
     }
@@ -190,7 +190,7 @@ void particleUtilsPS::setParticleSetTrailType(particleSetPS &particleSet, uint8_
 //does setParticleSetTrailRand(), but for a whole set
 void particleUtilsPS::setAllTrailRand(particleSetPS &particleSet, bool noTrails,
                                       bool oneTrail, bool twoTrail, bool revTrail, bool infTrail) {
-    for( uint8_t i = 0; i < particleSet.length; i++ ) {
+    for( uint16_t i = 0; i < particleSet.length; i++ ) {
         setTrailRand(particleSet, i, noTrails, oneTrail, twoTrail, revTrail, infTrail);
     }
 }
@@ -203,7 +203,7 @@ void particleUtilsPS::setAllTrailRand(particleSetPS &particleSet, bool noTrails,
 //  twoTrail -- trail type 2
 //  revTrail -- trail type 3
 //  infTrail -- trail type 4
-void particleUtilsPS::setTrailRand(particleSetPS &particleSet, uint8_t partNum, bool noTrails,
+void particleUtilsPS::setTrailRand(particleSetPS &particleSet, uint16_t partNum, bool noTrails,
                                    bool oneTrail, bool twoTrail, bool revTrail, bool infTrail) {
     uint8_t trailType = 0;
     //booleans = 1 for true and 0 for false, so we will randomly pick from as many true flags as we have
@@ -239,7 +239,7 @@ void particleUtilsPS::setTrailRand(particleSetPS &particleSet, uint8_t partNum, 
 //sets the length of a particle's trail
 //The trail length is varied by a random amount no greater than the passed in range
 //Trails must have a minimum length of 1
-void particleUtilsPS::setParticleSetTrailSize(particleSetPS &particleSet, uint8_t partNum, uint8_t trailSize, uint8_t range) {
+void particleUtilsPS::setParticleSetTrailSize(particleSetPS &particleSet, uint16_t partNum, uint8_t trailSize, uint8_t range) {
     if( trailSize == 0 ) {
         trailSize = 1;
     }
@@ -248,7 +248,7 @@ void particleUtilsPS::setParticleSetTrailSize(particleSetPS &particleSet, uint8_
 
 //sets a particle's bounce behavior on the passed in bounce value
 //to set the bounce randomly, pass in a value > 1
-void particleUtilsPS::setParticleSetBounce(particleSetPS &particleSet, uint8_t partNum, uint8_t bounce) {
+void particleUtilsPS::setParticleSetBounce(particleSetPS &particleSet, uint16_t partNum, uint8_t bounce) {
     if( bounce > 1 ) {
         bounce = random8(2);
     }
@@ -258,7 +258,7 @@ void particleUtilsPS::setParticleSetBounce(particleSetPS &particleSet, uint8_t p
 //sets a particle's colorIndex to the passed in value
 //If you want the color to be randomly picked from a palette
 //pass in the palette's length as colorIndex and set randColor as true
-void particleUtilsPS::setParticleSetColor(particleSetPS &particleSet, uint8_t partNum, uint8_t colorIndex, bool randColor) {
+void particleUtilsPS::setParticleSetColor(particleSetPS &particleSet, uint16_t partNum, uint8_t colorIndex, bool randColor) {
     if( randColor ) {
         colorIndex = random8(colorIndex);
     }
@@ -282,14 +282,14 @@ void particleUtilsPS::freeParticleSet(particleSetPS &particleSet) {
 //This should be the actual memory size of the array.
 //!!!DO NOT call this if the particle set was not created using malloc() or buildParticleSet()
 void particleUtilsPS::freeAllParticles(particleSetPS &particleSet) {
-    for( uint8_t i = 0; i < particleSet.maxLength; i++ ) {
+    for( uint16_t i = 0; i < particleSet.maxLength; i++ ) {
         freeParticle(particleSet, i);
     }
 }
 
 //Frees the pointer to a particle in a particleSet, should only be used if the particles was generated using malloc()
 //!!!DO NOT call this if the particle set was not created using malloc() or buildParticleSet()
-void particleUtilsPS::freeParticle(particleSetPS &particleSet, uint8_t partNum) {
+void particleUtilsPS::freeParticle(particleSetPS &particleSet, uint16_t partNum) {
     free(particleSet.particleArr[partNum]);
 }
